requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asttokens==3.0.0
    # via stack-data
bleach==6.2.0
    # via kaggle
build==1.2.2.post1
    # via pip-tools
certifi==2024.12.14
    # via
    #   kaggle
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   opendatasets
    #   pip-tools
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.1.1
    # via ipython
executing==2.1.0
    # via stack-data
fonttools==4.55.3
    # via matplotlib
idna==3.10
    # via requests
imbalanced-learn==0.13.0
    # via -r requirements.in
ipython==8.30.0
    # via -r requirements.in
jedi==0.19.2
    # via ipython
joblib==1.4.2
    # via
    #   -r requirements.in
    #   imbalanced-learn
    #   scikit-learn
kaggle==1.6.17
    # via opendatasets
kiwisolver==1.4.7
    # via matplotlib
matplotlib==3.9.3
    # via
    #   -r requirements.in
    #   seaborn
matplotlib-inline==0.1.7
    # via ipython
mpmath==1.3.0
    # via sympy
numpy==2.2.0
    # via
    #   -r requirements.in
    #   contourpy
    #   imbalanced-learn
    #   matplotlib
    #   pandas
    #   patsy
    #   scikit-learn
    #   scipy
    #   seaborn
    #   statsmodels
opendatasets==0.1.22
    # via -r requirements.in
packaging==24.2
    # via
    #   build
    #   matplotlib
    #   plotly
    #   statsmodels
pandas==2.2.3
    # via
    #   -r requirements.in
    #   seaborn
    #   statsmodels
parso==0.8.4
    # via jedi
patsy==1.0.1
    # via statsmodels
pexpect==4.9.0
    # via ipython
pillow==11.0.0
    # via matplotlib
pip-tools==7.4.1
    # via -r requirements.in
plotly==5.24.1
    # via -r requirements.in
prompt-toolkit==3.0.48
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==19.0.0
    # via -r requirements.in
pygments==2.18.0
    # via ipython
pyparsing==3.2.0
    # via matplotlib
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   kaggle
    #   matplotlib
    #   pandas
python-slugify==8.0.4
    # via kaggle
pytz==2024.2
    # via pandas
requests==2.32.3
    # via kaggle
scikit-learn==1.6.1
    # via
    #   -r requirements.in
    #   imbalanced-learn
    #   sklearn-compat
scipy==1.14.1
    # via
    #   -r requirements.in
    #   imbalanced-learn
    #   scikit-learn
    #   statsmodels
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via
    #   kaggle
    #   python-dateutil
sklearn-compat==0.1.3
    # via imbalanced-learn
stack-data==0.6.3
    # via ipython
statsmodels==0.14.4
    # via -r requirements.in
sympy==1.13.3
    # via -r requirements.in
tenacity==9.0.0
    # via plotly
text-unidecode==1.3
    # via python-slugify
threadpoolctl==3.5.0
    # via
    #   imbalanced-learn
    #   scikit-learn
tqdm==4.67.1
    # via
    #   kaggle
    #   opendatasets
traceml==1.2.0
    # via -r requirements.in
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via
    #   kaggle
    #   requests
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
